Added level 1 and fixed save/load data management

+++ Grapple Ball/Assets/_Scripts/PlayerData.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 [Serializable]
 public class PlayerData
 {
-	public int playerHealth;
+	public float playerHealth;
 	public int playerGemCount;
 
 	public PlayerData(PlayerStatus player) {
